fix(prerender): static metadata route should be optimized by default (#2950)
