+++ packages/vinext/src/entries/app-rsc-entry.ts
@@ -1370,6 +1370,7 @@ const __appRscHandler = createAppRscHandler({
       isrSet: __isrSet,
       makeThenableParams,
       scheduleBackgroundRegeneration: __triggerBackgroundRegeneration,
+      consumeDynamicUsage,
     });
   },`
       : ""
